install/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
aiosqlite==0.20.0
alembic==1.13.1
python-dotenv==1.0.1
pydantic>=2.9.0
pydantic-settings>=2.2.1

# AI Provider SDKs (pinned to avoid resolution issues)
openai==1.50.0
anthropic==0.34.0
google-generativeai==0.7.0
ollama==0.4.0
tiktoken==0.7.0

# HTTP client
httpx>=0.27.0

# System monitoring
psutil==5.9.8

# File processing
pypdf>=5.0.0
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
Pillow==10.3.0
python-multipart>=0.0.18

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==4.1.0

# Development
black==24.3.0
flake8==7.0.0
mypy==1.9.0
